// pkg/packages/gcp/src/generated/run/model/googleLongrunningOperation.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cloud Run Admin API
 * OpenAPI spec version: v2
 */
import type { GoogleLongrunningOperationMetadata } from './googleLongrunningOperationMetadata';
import type { GoogleRpcStatus } from './googleRpcStatus';
import type { GoogleLongrunningOperationResponse } from './googleLongrunningOperationResponse';

/**
 * This resource represents a long-running operation that is the result of a network API call.
 */
export interface GoogleLongrunningOperation {
  /** The server-assigned name, which is only unique within the same service that originally returns it. If you use the default HTTP mapping, the `name` should be a resource name ending with `operations/{unique_id}`. */
  name?: string;
  /** Service-specific metadata associated with the operation. It typically contains progress information and common metadata such as create time. Some services might not provide such metadata. Any method that returns a long-running operation should document the metadata type, if any. */
  metadata?: GoogleLongrunningOperationMetadata;
  /** If the value is `false`, it means the operation is still in progress. If `true`, the operation is completed, and either `error` or `response` is available. */
  done?: boolean;
  error?: GoogleRpcStatus;
  /** The normal, successful response of the operation. If the original method returns no data on success, such as `Delete`, the response is `google.protobuf.Empty`. If the original method is standard `Get`/`Create`/`Update`, the response should be the resource. For other methods, the response should have the type `XxxResponse`, where `Xxx` is the original method name. For example, if the original method name is `TakeSnapshot()`, the inferred response type is `TakeSnapshotResponse`. */
  response?: GoogleLongrunningOperationResponse;
}